Write-Host "📋 Available files in $buildDir:" The error message indicates:
Variable reference is not valid. ':' was not followed by a valid variable name character. Consider using ${} to delimit the name.

The Problem
PowerShell is having trouble parsing the variable $buildDir when it's followed immediately by a colon :. PowerShell thinks the colon is part of the variable name.
The Fix
You need to use curly braces to properly delimit the variable name:
powershellWrite-Host "📋 Available files in ${buildDir}:"